Further restrict "weird" 0% columns
An old input file used a general-account product, but selected some
zero-percent columns that should be available only for separate-account
products. These weird columns were available because inforce separate
account value exceeded zero, which shouldn't be possible (and indeed
that field is grayed in lmi today, but presumably wasn't a decade ago
when this file was created). XSL-FO allowed an illustration to be
printed, although columns of nonexistent separate-account values were
all blank. However, the newer wxPdfDoc replacement threw an exception,
which is a reasonable behavior, yet one that end users would not
welcome. With this commit, the impermissible columns are removed from
the supplemental report, and unavailable in the GUI, for products that
don't offer a separate as well as a general account.