Fix an off-by-one error
Certain products have a crediting rate that increases annually, reaching
an "ultimate" value in the sixth year that doesn't change thereafter:
   v0, v1, v2, v3, v4, v5...v5
  1st 2nd 3rd 4th 5th 6th  6th
Along with the current year's rate, the ultimate rate is to be printed
whenever it differs from current. The old XSL template confusingly
expressed that condition as
  InforceYear &lt;= 4
i.e., we haven't reached the sixth year if the zero-based index is less
than or equal to four; [0,4] is more naturally expressed as [0,5). Then
the difference between "&lt;=" and '<' was lost in translation to
  add_variable("InforceYearLE4", inforce_year < 4)
which has now been replaced by
  add_variable("InforceYearLT5", inforce_year < 5)

Eventually this product-specific logic should be abstracted into the
product database: it never belonged in presentation code.