Facilitate dimensional analysis
In vim,
  /.*lines\&.*rows/
highlights potential dimensional-analysis errors here.

This is okay because the conversion is deliberate:
  int const lines_per_group = rows_per_group + 1;
and has been explained by adding a comment.

This is not obviously okay:
  if(rows_on_last_page <= free_lines)
but has been explained by adding an assertion: rows and lines are
comparable for a terminal group when that group is partial.