Replace certain local variables with data members
This begins a series of commits that demonstrates design tradeoffs.

This commit is desirable because it makes the results of intermediate
calculations available to new member functions that may be added later,
and, if const accessors are added, to clients of class paginator--yet
one must rue the loss of constness.