Make the just-added data members const
See the immediately preceding commit.

This commit recovers the constness of the penultimate revision, which
is great. And it initializes const members in the ctor-initializer,
which is fine (and necessary). But it introduces a latent defect--if
the newly-added unit test is not commented out, then this line:
    ,groups_per_page_    {(max_lines_per_page_ + 1) / lines_per_group_}
results in:
  Unhandled exception: divide by zero
That problem had previously been precluded by assertions in the body
of the ctor. Now, the assertions come too late: the body isn't reached
because a fault occurs in the ctor-initializer.