Remove the latent defect just added
See the two immediately preceding commits.

Abnormal termination is a grave defect, which this commit removes by
asserting the validity of ctor arguments at the proper moment. To see
why this option seems better than any other, consider the tradeoffs:

 - Constness could be forsaken.

 - Constness could be subverted with const_cast.

Or assertions could be made at the proper time by alternative means:

 - Push them into a base class, complexifying the design.

 - The comma operator won't help, but the conditional operator could
   be used in the ctor-initializer, e.g.:
     ,rows_per_group_ {0 < rows_per_group ? rows_per_group : throw 0;}
   which is less readable than this commit.

 - Templates. SFINAE. A barrier to comprehension: overkill here.

 - The assertions could be expressed in an macro. Not.

 - The possibility of abnormal termination could be disregarded,
   because how could that go wrong?

Join these two lines to produce one URL:
  https://isocpp.org/wiki/faq/ctors
  #initializing-members-from-other-members
for a discussion of the morality of this commit, which warns of "subtle
order-dependency errors if someone reorganizes the layout of member
objects". Yet these calculations are manifestly order-dependent, and
rearranging the lines of code in the antepenultimate commit would have
the same effect as rearranging the mem-initializers here. The
precondition-assertion function could be rewritten to reference the ctor
arguments directly, e.g.:
   foo::foo(int a, int b)
   :ctor_args_are_sane_ {assert_preconditions(a, b)}
   ,[all other mem-initializers]
which would arguably alleviate the member-object order dependency, but
at the cost of creating a dependency on the order of those arguments:
   :ctor_args_are_sane_ {assert_preconditions(b, a)} // Oops!
while also introducing obscurity and maintenance overhead.