Rework the commit just reverted
See the four immediately preceding commits.

Upon reflection, this part of the penultimate commit message:
|     ,rows_per_group_ {0 < rows_per_group ? rows_per_group : throw 0;}
|   which is less readable than this commit.
seems incorrect: "less" should have been "more". Accordingly, used that
technique in this commit, on the only argument for which a value of
zero could cause a division-by-zero fault. See:
  https://lists.nongnu.org/archive/html/lmi/2018-09/msg00001.html

Now that the ctor arguments are less obviously equivalent to their
corresponding data members, restated the precondition assertions in
terms of the arguments themselves.