Rewrite the paginator logic so that it has a single exit point
Generalized the logic in two places so that it's correct even for zero
rows of data. Moved special handling for the zero-rows case to the exit
point: printing one empty page in this case is properly an optional
behavior rather than an inherent part of the design, and someday it
might be desirable instead to handle that special case outside this
function.