Avoid guessing initial values
When a member ought to be initialized (e.g., here, where all members are
handled in the ctor-initializer), but its value cannot be determined at
the point of initialization, prefer '{}' to guessing a value.

This former initial value has the dubious rationale that variables with
similar names might have the same value:
-    ,lines_on_last_page_ {lines_on_full_page_}
and this one had been set explicitly in an early-exit conditional that
has been removed:
-    ,page_count_         {1}

An alternative idea, not used here because it would be overkill, would
be to choose a deliberately implausible value as a "canary", e.g.,
    ,page_count_         {INT_MIN}
and, at exit, assert as a postcondition that the actual value is in a
plausible range that excludes that initial value.