Invert control for unit test
Replaced free function paginate() with new class paginate, whose member
format() duplicates the control structure of the original while virtuals
implement the incidental details.

This emerging design is not yet ideal. Names for classes with distinct
responsibilities could be more distinct than 'paginator' and 'paginate';
but it is not yet known whether they should even be distinct classes.
It remains to be seen whether the virtuals are well chosen, and whether
they should be const. While class paginator is already used in lmi, for
the moment new class paginate is confined to a unit test where radical
changes are still easily made.