Forestall latent off-by-one errors
In the section changed, 'year' was in index origin one, but everywhere
else it is in origin zero. Thus, in a loop over 'year', the counter
was asserted to lie within (0, size()] , and C++ arrays were actually
indexed by 'year-1'. It did do the right thing, but a consistent index
origin makes the revised version easier to understand and maintain.