Improve documentation
"mostly": In the general case, a virtual page is rendered across several
physical pages. The special case where a single page suffices is less
common that it may at first have appeared.

"// Make base class dtor virtual.": This could be read as an author's
note to himself, meaning "don't forget to make this class's base class's
dtor virtual." Apparently it means "this class is intended to be derived
from, so its dtor is virtual", which doesn't really need a comment.