Embark upon inversion of control by making table generator a member
The goal of this and the next few commits is to invert control in
rendering paginated tables. Present substeps in render() will become
overrides; yet the pagination base class naturally knows nothing about
the necessary local state; and therefore the local state variables of
render() must become data members of the class. This first commit lays
that groundwork with respect to the table generator, which is now,
incidentally, instantiated OAOO in this class.