Proceed toward IoC by holding ubiquitous arguments as members
See the immediately preceding commit message. The ledger and pdf_writer
arguments that are passed to so many functions cannot be passed to
overrides of class paginator's virtuals, which take no arguments.

Arguably those arguments should be members of root class logical_page,
with protected accessors. The alternative, here, is moderately long-
distance friendship and a couple of const_casts. Another option, worse
than either of those, would be
  virtual paginator::whatever(void* blob)
but that's mentioned only for completeness.