Invert control of pagination
This completes a series of three consecutive commits that have moved
the pagination control structure out of a GUI module and into a pure C++
module with extensive unit tests.

Incidentally, this resolves one of the last major discrepancies between
this code and its XSL-FO predecessor, viz.: an illustration that lapses
before the first anniversary now has both headers and footers on the
otherwise-blank pages with zero rows of data--whereas, until now, only
the footers were printed.