Print "0%" in all zero-percent column headers
Added "0%" to two headers. The omission for the first two general-
account columns was obviously a mistake, as "0%" is included for the
corresponding separate-account columns a couple lines below. This
mistake was copied faithfully from the 'ledger_xml_io.cpp' original
that is used for XSL-FO, where it is left unfixed because the XSL-FO
implementation is about to be expunged.