Add newlines to zero-percent column headers
Added a newline after "Curr" or "Guar" and before "Charges". The XSL-FO
original does this implicitly, so this change removes a gratuitous
difference between old and new PDF implementations. More importantly, it
makes the headers of these columns narrower so that a full set of zero-
percent columns can fit in one supplemental report.