Add a yearly-varying fixed loan rate to a (renamed) exotic product
Renamed 'sample2quo' to 'sample2xyz', because it can serve exotic
purposes other than weird "split-premium" group quotes. Added another
exotic feature--a varying fixed loan rate--in order to demonstrate
a problem (soon).

Note on argot: loan rates may be either "fixed" (fully specified in the
contract when it is issued) or "variable" (declared from time, with
reference to a published financial index, as prescribed by regulation).
Thus, if the contract specifies a rate of 6% for the first policy year
and 5% thereafter, that is a "fixed" rate: although it varies by year,
it is emphatically not a "variable loan rate" as defined by regulation.

Similarly, in C++, this is a type-specifier:
  bool
and this:
  bool // Boolean type.
is the same type-specifier, elaborated with a comment, so it's an
elaborated type-specifier, but it is not an elaborated-type-specifier.