Change the format of 'GenAcctAllocation'
Changed the format of this field to a percentage, scaled by 100. It's a
proportion, so the original zero-decimals format was not very useful. No
corresponding change is made to 'ledger_xml_io.cpp', the corresponding
file for XSL-FO: it is better not to touch stable code that is soon to
be expunged, and XSL templates override this format anyway.