Favor problem-domain terminology
A question like "what's the premium load?" would immediately be
answered with something like "three percent". A question like "what
are the premium loads?" would more likely draw a response such as
"two percent for premium tax, and one percent for DAC tax". Here,
the former is wanted. It's just unfortunate that the XSL-FO original
used the latter, yet at this moment the goal is to reproduce the old
system's PDF output, not to improve it; but internal C++ variable
names should be as good and consistent as possible.

It is especially problematic to use a name such as "admin charges",
which looks like a term of art but lacks any universal definition.
It might be defined in different ways by different contracts. The
replacement "misc charges" more clearly cannot be a universal term:
a deliberately uninformative name is better than a misleading one;
being forced to ask is better than being invited to assume, wrongly.