Do not depict a scalar as a vector
The rate of interest due on loans can vary by year. The old XSL-FO
system depicted it in a column that actually suppressed any variation,
replicating the initial value instead. That error is fixed by this
commit, only for the new PDF code (the old will soon be expunged). See:
  https://lists.nongnu.org/archive/html/lmi/2018-02/msg00100.html