Avoid assertions due to invalid font during PDF generation
Don't pass invalid font to wxDCFontChanger, in some circumstances it can
be used before it's changed to something else, so use the default font
instead.

Another possibility would be to stop using wxDCFontChanger entirely and
replace it with some (non-existent) wxDCFontPreserver that would just
restore the original font in its dtor but wouldn't change the font in
its ctor at all. However it's not clear if introducing a separate class
is really worth it, so for now continue using wxDCFontChanger.