Refactor
Class ledger_evaluator contains all the data written with
  --pyx=values_tsv
so it's reasonable to put that code in a member function of that class.

The class Ledger instance is passed to this new member function because
a couple of Ledger members are used, at least for now.

The PDF output filename is also passed, although it's not yet used. It
will soon be used instead of the static constant filename 'values.tsv',
which unreasonably limits the usefulness of this facility.