In TSV, show scalars' names and values in two columns only
"Vectors" are shown thus:
  name0    name1   ...
  value00  value10 ...
  value01  value11 ...
  ...      ...     ...
which is natural because their values are almost all numeric with no
more than about ten digits. "Scalars" were shown similarly:
  name0    name1  ...
  value0   value1 ...
as an experiment, removed in this commit because it clearly failed:
it was unnatural because their values are often strings of arbitrary
length. Now scalars are shown only thus:
  name0   value0
  name1   value1-which-may-sprawl-all-the-way-across-the-screen
  ...     ...
Revised a comment block accordingly. Incidentally, rearranged the
code for clarity.