Fix initial font after detaching the header HTML cell
Don't detach the initial wxHtmlColourCells and wxHtmlFontCell when
removing the header from the HTML document as this resulted in using the
default back- and foreground colours and font when rendering the part of
the document without the header later, which was wrong as it didn't use
the font explicitly set for all the HTML contents by pdf_writer_wx.

In practice, this resulted in the cells following <header> tag using a
visible wrong font until the next font change, e.g. a <b>, <i> or <u>
tag occurrence.

The fix itself is trivially simple and consists in only including the
header contents itself in the header cell container, excluding the
contents of the currently opened container which was erroneously
included before.