Reorder a statement
In this group loop, it's necessary only to follow the same order as
IllusVal::run():
  - create an AccountValue object
  - set its monthly-trace filename
  - perform the same steps as AccountValue::RunAV()
Moving the statement that stores the AccountValue object in a container
has no effect on observable results.

However, the code was not robust:
  smart_pointer p(values);
  p->do_something();
  container.push_back(p);
  p->do_something_else();
because it works with std::shared_ptr, but not with std::unique_ptr
after the necessary change:
  p->do_something();
- container.push_back(p);
+ container.push_back(std::move(p));
  p->do_something_else(); // oops: p has been moved from
The next commit, separated for clarity, further improves robustness.