Replace std::shared_ptr with std::unique_ptr
This change brings no great benefit per se, though it does demonstrate
the usefulness of the refactoring spread over the last several commits.
Even better than replacing a heavy smart pointer with a light one will
be to replace a smart pointer with the object it would hold: that's
simpler, and gets rid of the problem that pointers can be null.