Don't add extra page if there are no more rows in group quotes
Check if there is enough space left on the current page before
outputting a new row, as we can be certain that we do need a new page in
this case.

Previous code had a bug when the total number of rows was exactly equal
to the number fitting on the first page as it added a superfluous page
in this case, resulting in an assertion failure later and wrong output.