Eliminate an unwanted blank line
Rewrote patterns such as this:
  {{#Condition}} abc {{/Condition}}
  {{^Condition}} def {{/Condition}}
  <br>
in this way:
  {{#Condition}} abc<br> {{/Condition}}
  {{^Condition}} def<br> {{/Condition}}
as is generally done in other '*header*.mst' files. It is more robust to
write the <br> element explicitly in this manner instead of factoring it
out of the conditionals: in one case, this usage:
  {{^IsInforce}}
    Conditional text with no "{{#" alternative.
  {{/IsInforce}}
  <br>
resulted in an unwanted blank line in the "{{#" case.