Fix initializing a new HTML document containing the given cell
make_html_from() function didn't do anything useful as it made a shallow
copy of the new document it created (leaking the original one in the
process) instead of simply taking ownership of it.

This meant that the object returned from this function did not contain
the color and font cells which was the sole reason for its existence in
the first place and, due to this, the header cells created using this
function didn't use the correct initial font, resulting in visibly wrong
appearance, which is fixed by this change.