Add and test product_database copy ctor
Class BasicValues is movable, and is to remain so. Its
  std::shared_ptr<product_database>   Database_;
member is to be replaced with a plain class instance. For this
particular replacement:
  product_database database_;
the product_database move ctor added two commits ago (commit 64a9cebf)
would suffice, but this preferred alternative
  product_database const database_;
doesn't compile with the default move ctor--at least with gcc-7.3,
a copy ctor is required.