Replace ...cast<T>(...Query(...)) with ...Query<T>(...)
After these changes, this command
  grep 'cast.*Query' *.?pp
finds only
  ihs_basicval.cpp: ,bourn_cast<int>(database().Query(TableID))
  ihs_basicval.cpp: ,bourn_cast<int>(database().Query(TableID, index))
The second line uses a nondefault 'index', and no function template is
supplied for that extraordinary use case. The first line cleaves to its
original form for concinnity with the second.

These changes are at least terser (if bourn_cast had been used), and
also safer (if static_cast had been used). They still aren't ideal
because in a typical case like this:
-    maturity_age_ = static_cast<int>(Query(DB_MaturityAge));
+    maturity_age_ = query<int>(DB_MaturityAge);
either 'maturity_age_' is of type int and the explicit template
parameter is redundant, or it is not and the explicit template
parameter is incorrect. Later commits will address this, thus e.g.:
-    maturity_age_ = query<int>(DB_MaturityAge);
+    query_into<int>(DB_MaturityAge, maturity_age_);
transitionally, and eventually thus:
-    query_into<int>(DB_MaturityAge, maturity_age_);
+    query_into(DB_MaturityAge, maturity_age_);