Make a deliberate mistake more obvious in a unit test
Treating DB_MaturityAge as type oenum_alb_or_anb looked plausibly
accidental, because it is not altogether unreasonable to say that
"maturity age" is "age last birthday". Treating DB_ChildRiderMinAmt
as that type is more easily seen as deliberately mistaken: to say
that a rider's minimum amount is "age last birthday" is clearly
nonsensical.