Finish DBO renaming: risky treatment of initial premium
Replaced
  Initial7702BftIsDb
with
  RiskyInitial7702Db
because the former is ambiguous (it sounds like it should always be
true for the MDB DBO), while the second emphasizes the real danger of
treating anything higher than the initial specamt as the 7702 death
benefit--contrast section 3/1 here:
  https://www.nongnu.org/lmi/7702.html
This risky option isn't implemented anyway. For a contract issued thus:
  DBO: increasing (or perhaps ROP)
  100,000 specamt
    7,000 net premium with application
the idea is to set the specamt at 100K, but treat the 7702 death benefit
as 107K, exposing the purchaser to adverse tax consequences should poor
investment performance cause the actual death benefit to decrease.

This concludes the renaming of DBO-related database entities.
'AllowChangeToDbo2' will not be renamed; it has been marked as
obsolescent, and will be expunged ere long.