Integrate Boost into MSVS solution
Create MSVS projects for building the libraries we use and include them
in the solution file.

Also use a properties file setting up everything Boost-related in a
single place instead of scattering the options needed for building Boost
everywhere.

This ensures that once Boost dependency is removed, we can just remove
the property file and all the references to it without changing anything
else.