Improve gcc dialect defaults
Formerly, C++1998 was specified as the default dialect, and then
overridden for each successive gcc version since 2005. Now, the
current dialect is the default. This make more sense because
 - it makes upgrading to a later gcc version easier; and
 - lmi actually does require the C++2017 dialect or later.

Because contemporary lmi code can't even be built with older tools
such as 2005's gcc-3.4.5, the overrides for such older tools are
retained for historical reasons only; clearly C++2017 code cannot be
built with a pre-C++2017 compiler, no matter what options are given.
But it is useful to keep such history in current makefiles--e.g., a
comment like:
  # Use a correct snprintf() implementation:
  #   http://article.gmane.org/gmane.comp.gnu.mingw.user/27539
  cxx_standard += -posix
may come in handy someday if a problem with snprintf() arises and
grepping the source for that function's name finds that comment readily.
(Commands like 'git-grep', 'git-log --grep', and 'git-log -G' would of
course find it even if it were deleted now, so this tradeoff accepts a
little clutter in return for making the simpler command just work.)