Print a page number on the duplicate signature page
NAIC drafted its illustration reg two decades ago, when illustrations
were normally printed on paper. At that time, it made sense to print an
extra copy of the prescribed numeric-summary-and-certification page, so
that it could be signed and physically delivered to the insurer.

Prior to this commit, that duplicate page has borne no page number, yet
it has been included in the page count; thus, an illustration could
consist of a (questionably) unnumbered cover page, then eight pages
numbered "1 of 9" through "8 of 9"...and finally this unnumbered
duplicate page, so that no page "9 of 9" exists and the illustration
therefore seems incomplete. Now, the last page is numbered "9 of 9".

Excluding the cover page, the now-retired XSL-FO code would have
numbered the pages "1 of 8" through "8 of 8" in the example above, and
labelled the duplicate final page "Attachment". It would alternatively
have been okay to use that labelling scheme with wxPdfDoc illustrations,
but that would constitute a major change. At first glance, it may appear
that former class ill_reg_numeric_summary_attachment could simply be
derived from class page_with_marginals (because it needs headers and
footers), but then it wouldn't support an MST template, which requires
derivation from class standard_page...which in turn derives from class
numbered_page, but all numbered_page instances participate in the total
page count.