Add "split funds" supplemental report to FINRA illustrations
When SplitFundAllocation is true, automatically include the predefined
"split funds" report, which is similar to a supplemental report, but
with the columns being fixed (albeit fixed slightly differently
depending on whether SplitMinPrem is true or not) instead of being
selected by the user.

Also, this report is always generated unconditionally for illustrations
using split funds and doesn't have to be explicitly requested.