Expunge a useless conditional
It seems natural to suppose that 'ProducerCity' could be an empty
string. However, it's actually the concatenation of several strings,
one of which (Input::AgentState) can never be empty. At any rate,
the removed conditional
  ProducerCity != "0"
could never be true, because Input::AgentState is always two characters
long, so a string containing it can never match a one-character string
consisting of the numeral zero. (Probably something like "length != 0"
was intended, but that could never be true, either.)