Replace initial with maximum loan due rate in ledger
Reports use this value only in contexts where a scalar is portrayed as
representing a vector whose elements actually may vary from year to
year--a purpose for which the maximum is better suited because this
portrayal is in the nature of a guarantee.

The corresponding spread is already "guaranteed not to exceed" a given
(defectively hardcoded) value as depicted in 'ill_reg_narr_summary.mst'.
Both the gross rate and the spread may vary by year; if, for brevity,
the year-by-year values aren't to be shown, then the (scalar) maximum
should be used for both, and composites should use the maximum value
across all cells.