Fix defect introduced 20160309T0113Z: weird policy-name conditional
Added a new database field to control whether guaranteed-issue contracts
in Texas should be captioned as substandard.

It is unclear what defect was actually fixed on 20160309T0113Z with
commit message:
  Fix defect introduced 20050720T1018Z: "substandard standard"
because TX GI contracts with standard rates still print the antinomical
"substandard standard" caption. That 2016 revision may simply have
restricted the antinomy, in practical effect, to a single product--and
that is the best that can be done unless the word "standard" can be
omitted altogether.