Make CLI and GUI getopt code more similar
Removed pointless boolean flags. If an option performs an imperative
action, it's simpler to inline that action's code in the option handler.

Removed an accidental '-b' option. It was apparently just a placeholder
in the original implementation, which had never been removed until now.

Made all options that are documented "...and exit" actually return (some
had not done so). Immediate-return options return 0 in the CLI; in the
GUI, they return nonzero, merely because that's a convenient way to tell
wx to exit.