Rename CSO table module, part 1 of 2
Made module name identical to the name of the function it provides, so
that searching for the name of the header in an #include directive
(e.g., with '*' in vim) finds calls to that function.

This part 1 commits only the result of 'git mv', separately to make the
history as clear as possible. It has been committed with '--no-verify',
because it fails the concinnity-test hook. It will be immediately
followed by a commit to fix the build.