Allow building with gcc-8.2.0
As to the warnings suppressed for gcc-8.2.0, see:
  https://lists.nongnu.org/archive/html/lmi/2019-03/msg00006.html

As to the behavior of '-Wno-*', see:

  https://gcc.gnu.org/onlinedocs/gcc/Warning-Options.html
| no diagnostic is produced for -Wno-unknown-warning unless other
| diagnostics are being produced

Perhaps this justifies the behavior documented in this revision:
'-Wno-noexcept' and '-Wno-useless-cast' are known to g++, so arguably
they aren't "unknown" to gcc, although it's unclear why a diagnostic
is useful for any '-Wno-' option. Of course, suppressing these
diagnostics is a workaround, and the most appropriate action is to
eradicate the root causes.