Change the line that provokes gcc-8.x '-Wuseless-cast' diagnostics
Ostensibly because of this updated boost patch, the same diagnostic need
no longer be suppressed for one list of boost sources in the makefile,
even for gcc-7.x .

To verify that all vestiges of recent temporary workarounds have been
removed, run this command:
  git diff f4c47d6e THIS-SHA1 workhorse.make
substituting this commit's SHA1 where indicated.

To apply the patch and rebuild, use the applicable commands given in
'install_mingw.make', viz.:

  make $coefficiency --output-sync=recurse -f install_miscellanea.make \
   clobber

  make $coefficiency --output-sync=recurse -f install_miscellanea.make

then rebuild lmi as usual.