Refactor: specify makefile directories at a lower level
The goal is to replace literal occurrences of /opt/lmi/local with a
flexible directory structure for multiple-architecture builds--see:
  https://lists.nongnu.org/archive/html/lmi/2017-05/msg00034.html

This commit is just a preliminary step. Specifying all standard (and
some nonstandard) directories in a low-level makefile just makes sense,
and it's desirable for the particular reason that 'GNUmakefile' already
uses what is now named $localbindir.

It is laudable that git's '--color-moved=plain' option finds blocks
moved between files.