Note that DB_MinPmt is not implemented
Products may prescribe a minimum payment amount in order to disallow
tiny payments (of one cent, e.g.) that aren't worth the expense of
accepting. This makes sense for an admin system that allows home office
staff the discretion to override such limits (although the current trend
is to replace clerical jobs with automation). A case can be made that
illustrations shouldn't depict any payment that actually would not be
accepted, in order to avoid creating a false expectation.

On the other hand, illustration systems don't allow human judgement, and
therefore must allow either all exceptions, or none. And some exceptions
really are reasonable. Suppose, e.g., that the lowest payment required
to prevent lapse exceeds the guideline limit. If the product defines a
minimum payment that's higher than the IRC 7702(f)(6) exception allows,
then the minimum will certainly be overridden. Twenty-first-century
insurers will need to codify such exceptions in their business rules.
Simplistic old-style specifications are not to be taken literally.