Detect build architecture correctly
$LMI_HOST had historically been defined for GNU/Linux cross builds
only, but undefined for cygwin, and had expediently but inappropriately
been used as a proxy for build type; 'uname' had served a similar role
elsewhere. Now 'config.guess' is used instead, throughout, by calling
  /usr/share/libtool/build-aux/config.guess
directly. That script is guaranteed to exist because 'libtool' is an
lmi dependency, and is assumed to be reasonably current. It potentially
compiles test code, so experience might show it to be too slow in some
circumstance; that doesn't seem to be a problem for debian GNU/Linux.

Third-party libraries now use their own 'config.guess' scripts (not
libtool's, and certainly not $LMI_HOST or 'uname').

* install_msw.sh: Set $platform conditionally to "Cygwin" precisely
because that string does not match 'uname' or 'config.guess' output.