Update PETE for gcc-8.x
Rebuilt PETE, thus:
  cd /opt/lmi/src/lmi/tools/pete-2.1.1
  make -f Makefile
  make -f Makefile distclean
Reverted files that changed only due to RCS expansion.

No one else need rebuild the library: lmi uses its generated files,
but none of them changed. It is however important to rebuild it from
time to time, just to ensure that it can be rebuilt.

Added a header to one file--it's unclear how PETE ever could have been
built without it. This was the smallest change that could suffice. The
headers are not idempotent, so the order of inclusion is significant.

* tools/pete-2.1.1/Makefile: Added a 'distclean' target. Removed a line
  in the 'clean' recipe that was intended to remove backup files.
* tools/pete-2.1.1/PETE/Tools/MakeOperators.cpp: Added a header that has
  become necessary.