Relocate lmi's build directory
This is the first in a series of changes to the directory structure,
all of which may require rebuilding.

Formerly, object files for a default build were written in
  /opt/lmi/src/build/lmi/gcc/ship/
but now they're instead written in
  /opt/lmi/gcc/build/ship/

Rationale: the old build directory
$(srcdir)/../build/$(notdir $(srcdir))/$(uname)/$(toolset)/$(build_type)
inserted 'build/' between the last two components of $(srcdir) (although
$(exec_prefix) would have been a better choice), then added $(uname)
(which recently became empty), and then added the rest. The new build
directory
  $(exec_prefix)/$(toolset)/build/$(build_type)
is rooted properly at $(exec_prefix); it includes $(toolset) before
the other path components because the existing $(toolset) and a host
type (soon to be added) will become part of $(exec_prefix), fulfilling
its destiny as the "parent directory for machine-specific binaries".