Prepare to number cover pages
Eventually, option (D) here:
  https://lists.nongnu.org/archive/html/lmi/2018-11/msg00045.html
will be wanted for all ledger types--though for the moment a legacy
version of the original class is retained under a different name in
order to postpone this change for 'finra' illustrations. For option (D),
the cover page class's place in the inheritance tree needs to change,
requiring a change in its location in the code. This preparatory commit
encompasses a block copy of an entire small class, along with changes
that can trivially be reviewed with '--color-moved=plain' (which doesn't
detect the block copy, unfortunately). Changing the order of the
add<cover_page> and start_numbering() lines has no effect yet because
class cover_page is not yet a numbered_page; and, as explained here:
  https://lists.nongnu.org/archive/html/lmi/2019-04/msg00035.html
it actually might not matter at all.