Don't eviscerate $(bindir)
'tabs/5/startup_script' assumes that it's safe to
  cd /opt/lmi/bin
when opening a terminal, and stay there in order to run programs
contained in that directory. Removing that directory in another
terminal causes...consequences. If it is used to store output files
in order to mimic the end-user experience, then those files should
not automatically be erased by any makefile target.