Rework MinGW gcc installation for cygwin
Merged the important parts of branch odd/fruitbat, which attempted to
make the 'INSTALL' instructions work for a cygwin-1.7 system installed
from crouchingtigerhiddenfruitbat.org .

* install_cygwin.bat: Repent of past practice (installing MinGW gcc in
the msw root, then using that installation through a cygwin mount),
which was cumbersome and unnecessary.

* install_mingw.make: Use only subdirectories of /opt/lmi for extraction
and installation of MinGW gcc archives. In the unfortunate case that the
msw-native '7-Zip' program is used, extracting to a directory outside of
cygwin can cause the extracted files and subdirectories to be invisible
to cygwin, and inaccessible, too--such that a cygwin administrative user
cannot 'rm' them due to msw ACLs. This may very well be the real reason
for the problem described here:
  https://lists.nongnu.org/archive/html/lmi/2018-03/msg00017.html
for which no convincing explanation had hitherto been found.
Incidentally, add another fallback '7-Zip' command in case the first one
fails, using 'cygpath' to translate its destination directory (as might
also be necessary for the first '7-Zip' command).

* install_libxml2_libxslt.make: Use new MinGW directories for cygwin.
* install_wx.sh: Likewise.
* install_wxpdfdoc.sh: Likewise.
* msw_cygwin.make: Likewise. Abandon the idea of installing multiple
MinGW gcc archives and selecting dynamically among them.