Copy a directory's contents instead of the directory as such
This change was present in branch odd/fruitbat, and should have been
included in commit 5f96fe41cc, but was overlooked.

Arguably '/path/to/.' might be preferable to '/path/to/*', here and
elsewhere throughout lmi. However, the perceived advantage is that
hidden files would be included, yet there should be no hidden files
here. And this version with '*' has already been tested with cygwin.